Add starter code for Node and Tree files
I've used a template repository for the simple fact that I had gemini create my eslint config file and it actually works now so instead of trying to figure that out at this time I'll reuse it here.
The last project used linked tree and node classes but instead of reusing Node I instead rewrote it.