Combine conditionals into one chain of events
In the delete method I for some reason had 3 if statements that were not chained together and those conditional's bodies did not have return statements so they would set a variable to a value and then keep executing the code in the else block which was leading to unwanted behavior and outcome. By changing this section into chaining these conditions it's not only easier to read but I didn't have to use return statements in each.