Change find when null result, add height method
Added the height method which returns the number of edges between a specific node with the argument value and its furthest leaf node. This method uses the find method and brought to light that I did not have a result for when a value is searched for but not found, so I've added in the optional chaining operator and an else condition to when checking the data of the node iterated into.